PY-105 Re-organize e2e tests
* tests/e2e/alpha - tests of the alpha API
* tests/e2e/old - tests of the old API
* tests/e2e/internal - tests of the internals
