+++ tests/e2e/conftest.py
@@ -1,45 +1,57 @@
-import logging
+import itertools as it
 import os
-import uuid
-from datetime import (
-    datetime,
-    timezone,
-)
+import time
+from concurrent.futures import Executor
+from datetime import timedelta
 
+import pytest
+from neptune_api import AuthenticatedClient
+from neptune_api.credentials import Credentials
 from neptune_scale import Run
-from pytest import fixture
 
-from neptune_fetcher import (
-    ReadOnlyProject,
-    ReadOnlyRun,
+from neptune_fetcher import ReadOnlyProject
+from neptune_fetcher.alpha.filters import Filter
+from neptune_fetcher.internal import identifiers
+from neptune_fetcher.internal.composition import concurrency
+from neptune_fetcher.internal.context import set_project
+from neptune_fetcher.internal.identifiers import RunIdentifier
+from neptune_fetcher.internal.retrieval.search import fetch_experiment_sys_attrs
+from neptune_fetcher.util import (
+    create_auth_api_client,
+    get_config_and_token_urls,
+)
+from tests.e2e.alpha.data import (
+    NOW,
+    PATH,
+    TEST_DATA,
 )
 
-# NOTE
-# The fixtures below assume that we're testing on a "many metrics" project
-# populated using the tests/populate_projects.py script, and that the project
-# contains ONLY data generated by the script.
-# If this is not the case, the tests will fail, as there are hard assumptions
-# around names and values of the metadata.
+API_TOKEN_ENV_NAME: str = "NEPTUNE_API_TOKEN"
 
 
-@fixture(scope="session", autouse=True)
-def cleanup_logging_handlers():
-    """Remove all logging handlers after each test session, to avoid
-    messy errors from the `logging` module.
+@pytest.fixture(scope="session")
+def client() -> AuthenticatedClient:
+    api_token = os.getenv(API_TOKEN_ENV_NAME)
+    credentials = Credentials.from_api_key(api_key=api_token)
+    config, token_urls = get_config_and_token_urls(credentials=credentials, proxies=None)
+    client = create_auth_api_client(
+        credentials=credentials, config=config, token_refreshing_urls=token_urls, proxies=None
+    )
 
-    The errors happen because `Run` installs `Run.close` as an `atexit` handler.
-    The method logs some messages. The output is captured by pytest, which closes
-    its logging handler early, causing "ValueError: I/O operation on closed file."
-    """
-
-    try:
-        yield
-    finally:
-        logger = logging.getLogger("neptune")
-        logger.handlers.clear()
+    return client
 
 
-@fixture(scope="module")
+@pytest.fixture(autouse=True)
+def context(project):
+    set_project(project.project_identifier)
+
+
+@pytest.fixture(scope="module")
+def executor() -> Executor:
+    return concurrency.create_thread_pool_executor()
+
+
+@pytest.fixture(scope="module")
 def project(request):
     # Assume the project name and API token are set in the environment using the standard
     # NEPTUNE_PROJECT and NEPTUNE_API_TOKEN variables.
@@ -53,117 +65,79 @@ def project(request):
     return ReadOnlyProject(project=project_name)
 
 
-@fixture
-def all_run_ids():
-    return sorted([f"id-run-{i + 1}" for i in range(6)])
+@pytest.fixture(scope="module")
+def run_with_attributes(project, client):
+    runs = {}
+    for experiment in TEST_DATA.experiments:
+        project_id = project.project_identifier
+
+        existing = next(
+            fetch_experiment_sys_attrs(
+                client,
+                identifiers.ProjectIdentifier(project_id),
+                Filter.name_in(experiment.name),
+            )
+        )
+        if existing.items:
+            continue
+
+        run = Run(
+            project=project_id,
+            run_id=experiment.run_id,
+            experiment_name=experiment.name,
+        )
+
+        run.log_configs(experiment.config)
+        # This is how neptune-scale allows setting string set values currently
+        run.log(tags_add=experiment.string_sets)
+
+        for step in range(len(experiment.float_series[f"{PATH}/metrics/step"])):
+            metrics_data = {path: values[step] for path, values in experiment.float_series.items()}
+            metrics_data[f"{PATH}/metrics/step"] = step
+            run.log_metrics(data=metrics_data, step=step, timestamp=NOW + timedelta(seconds=int(step)))
+
+            series_data = {path: values[step] for path, values in experiment.string_series.items()}
+            run.log_string_series(data=series_data, step=step, timestamp=NOW + timedelta(seconds=int(step)))
+
+        run.assign_files(experiment.files)
+
+        runs[experiment.name] = run
+    for run in runs.values():
+        run.close()
+
+    # Make sure all experiments are visible in the system before starting tests
+    for _ in range(15):
+        existing = next(
+            fetch_experiment_sys_attrs(
+                client,
+                identifiers.ProjectIdentifier(project.project_identifier),
+                Filter.name_in(*TEST_DATA.experiment_names),
+            )
+        )
+
+        if len(existing.items) == len(TEST_DATA.experiment_names):
+            return runs
+
+        time.sleep(1)
+
+    raise RuntimeError("Experiments did not appear in the system in time")
 
 
-@fixture
-def all_experiment_ids():
-    return sorted([f"id-exp-{i + 1}" for i in range(6)])
+@pytest.fixture(scope="module")
+def experiment_identifier(client, project, run_with_attributes) -> RunIdentifier:
+    from neptune_fetcher.alpha.filters import Filter
+    from neptune_fetcher.internal.retrieval.search import fetch_experiment_sys_attrs
+
+    project_identifier = project.project_identifier
+
+    experiment_filter = Filter.name_in(TEST_DATA.experiment_names[0])
+    experiment_attrs = extract_pages(
+        fetch_experiment_sys_attrs(client, project_identifier=project_identifier, filter_=experiment_filter)
+    )
+    sys_id = experiment_attrs[0].sys_id
+
+    return RunIdentifier(project_identifier, sys_id)
 
 
-@fixture
-def all_experiment_names():
-    return sorted([f"exp{i + 1}" for i in range(6)])
-
-
-@fixture
-def sys_columns_set(sys_columns):
-    return set(sys_columns)
-
-
-@fixture
-def sys_columns():
-    return ["sys/id", "sys/name", "sys/custom_run_id"]
-
-
-@fixture
-def id_to_name():
-    """Expected sys/custom_run_id -> sys/name"""
-    d = {f"id-run-{num}": "" for num in range(1, 7)}
-    d |= {f"id-exp-{num}": f"exp{num}" for num in range(1, 7)}
-
-    return d
-
-
-class SyncRun(Run):
-    """A neptune_scale.Run instance that waits for processing to complete
-    after each logging method call. This is useful for e2e tests, where we
-    usually want to wait for the data to be available before fetching it."""
-
-    def log(self, *args, **kwargs):
-        result = super().log(*args, **kwargs)
-        self.wait_for_processing()
-        return result
-
-    def log_configs(self, *args, **kwargs):
-        result = super().log_configs(*args, **kwargs)
-        self.wait_for_processing()
-        return result
-
-    def log_metrics(self, *args, **kwargs):
-        result = super().log_metrics(*args, **kwargs)
-        self.wait_for_processing()
-        return result
-
-    def add_tags(self, *args, **kwargs):
-        result = super().add_tags(*args, **kwargs)
-        self.wait_for_processing()
-        return result
-
-    def remove_tags(self, *args, **kwargs):
-        result = super().remove_tags(*args, **kwargs)
-        self.wait_for_processing()
-        return result
-
-
-@fixture(scope="module")
-def run_init_kwargs(project):
-    """Arguments to initialize a neptune_scale.Run instance"""
-
-    # TODO: if a test fails the run could be left in an indefinite state
-    #       Maybe we should just have it scoped 'function' and require passing
-    #       an existing run id
-    kwargs = {"project": project.project_identifier}
-    run_id = os.getenv("NEPTUNE_E2E_CUSTOM_RUN_ID")
-    if run_id is None:
-        run_id = str(uuid.uuid4())
-        kwargs["experiment_name"] = "pye2e-fetcher"
-    else:
-        kwargs["resume"] = True
-
-    kwargs["run_id"] = run_id
-
-    return kwargs
-
-
-@fixture(scope="module")
-def run(project, run_init_kwargs):
-    """Plain neptune_scale.Run instance. We're scoping it to "module", as it seems to be a
-    good compromise, mostly because of execution time."""
-
-    run = Run(**run_init_kwargs)
-    run.log_configs({"test_start_time": datetime.now(timezone.utc)})
-    run.wait_for_processing()
-
-    return run
-
-
-@fixture(scope="module")
-def sync_run(project, run_init_kwargs):
-    """Blocking run for logging data"""
-    return SyncRun(project=run_init_kwargs["project"], run_id=run_init_kwargs["run_id"], resume=True)
-
-
-@fixture
-def ro_run(project, run, run_init_kwargs):
-    """ReadOnlyRun pointing to the same run as the neptune_scale.Run"""
-    return ReadOnlyRun(read_only_project=project, custom_id=run_init_kwargs["run_id"])
-
-
-def pytest_set_filtered_exceptions() -> list[type[BaseException]]:
-    class DoNotFilterAnythingMarker(Exception):
-        pass
-
-    return [DoNotFilterAnythingMarker]
+def extract_pages(generator):
+    return list(it.chain.from_iterable(i.items for i in generator))
