+++ tests/e2e/internal/composition/test_attributes.py
