+++ tests/e2e/internal/composition/test_type_inference.py
